link apiCalls file to scripts.js
